全部更新新依赖

+++ app/src/main/java/com/winlator/ContainerFileManagerFragment.java
@@ -137,22 +137,22 @@ public class ContainerFileManagerFragment extends BaseFileManagerFragment<FileIn
     @Override
     public boolean onOptionsItemSelected(MenuItem menuItem) {
         int itemId = menuItem.getItemId();
-        switch (itemId) {
-            case R.id.menu_item_home:
-                folderStack.clear();
-                refreshContent();
-                return true;
-            case R.id.menu_item_view_style:
-                setViewStyle(viewStyle == ViewStyle.GRID ? ViewStyle.LIST : ViewStyle.GRID);
-                preferences.edit().putString("container_file_manager_view_style", viewStyle.name()).apply();
-                refreshViewStyleMenuItem(menuItem);
-                return true;
-            case R.id.menu_item_new_folder:
-                createFolder();
-                return true;
-            default:
-                return super.onOptionsItemSelected(menuItem);
+        if (itemId == R.id.menu_item_home) {
+            folderStack.clear();
+            refreshContent();
+            return true;
         }
+        else if (itemId == R.id.menu_item_view_style) {
+            setViewStyle(viewStyle == ViewStyle.GRID ? ViewStyle.LIST : ViewStyle.GRID);
+            preferences.edit().putString("container_file_manager_view_style", viewStyle.name()).apply();
+            refreshViewStyleMenuItem(menuItem);
+            return true;
+        }
+        else if (itemId == R.id.menu_item_new_folder) {
+            createFolder();
+            return true;
+        }
+        else return super.onOptionsItemSelected(menuItem);
     }
 
     private void setCurrentWorkingPath(String dosPath) {
@@ -343,28 +343,25 @@ public class ContainerFileManagerFragment extends BaseFileManagerFragment<FileIn
 
             listItemMenu.setOnMenuItemClickListener((menuItem) -> {
                 int itemId = menuItem.getItemId();
-                switch (itemId) {
-                    case R.id.menu_item_copy:
-                    case R.id.menu_item_cut:
-                        instantiateClipboard(file, itemId == R.id.menu_item_cut);
-                        break;
-                    case R.id.menu_item_remove:
-                        clearClipboard();
-                        ContentDialog.confirm(context, R.string.do_you_want_to_remove_this_file, () -> removeFile(file.toFile()));
-                        break;
-                    case R.id.menu_item_rename:
-                        clearClipboard();
-                        ContentDialog.prompt(context, R.string.rename, file.name, (newName) -> {
-                            file.renameTo(newName);
-                            refreshContent();
-                        });
-                        break;
-                    case R.id.menu_item_add_favorite:
-                        addFavorite(file);
-                        break;
-                    case R.id.menu_item_info:
-                        (new FileInfoDialog(context, file, container)).show();
-                        break;
+                if (itemId == R.id.menu_item_copy || itemId == R.id.menu_item_cut) {
+                    instantiateClipboard(file, itemId == R.id.menu_item_cut);
+                }
+                else if (itemId == R.id.menu_item_remove) {
+                    clearClipboard();
+                    ContentDialog.confirm(context, R.string.do_you_want_to_remove_this_file, () -> removeFile(file.toFile()));
+                }
+                else if (itemId == R.id.menu_item_rename) {
+                    clearClipboard();
+                    ContentDialog.prompt(context, R.string.rename, file.name, (newName) -> {
+                        file.renameTo(newName);
+                        refreshContent();
+                    });
+                }
+                else if (itemId == R.id.menu_item_add_favorite) {
+                    addFavorite(file);
+                }
+                else if (itemId == R.id.menu_item_info) {
+                    (new FileInfoDialog(context, file, container)).show();
                 }
                 return true;
             });
